+++ docs/migrate_from_openai_apps.md
@@ -1,5 +1,6 @@
 ---
 title: Migrate OpenAI App
+group: Getting Started
 description: Migrate from the OpenAI Apps SDK to MCP Apps SDK with concept mapping tables, API equivalents, and complete before/after code examples.
 ---
